+++ trip-itinerary-planner/libs/tripsphere/src/tripsphere/__init__.py
@@ -1,0 +1,2 @@
+"""Tripsphere protobuf package."""
+